refactor(idea_service): extract lifecycle/resonance views (#163)
  before: 1781
  after:  1579 (-202)
